[WIP] Add write github data to firestore
