+++ package.json
@@ -22,6 +22,7 @@
     "@slack/webhook": "^5.0.3",
     "axios": "^0.19.2",
     "dayjs": "^1.8.22",
+    "firebase-admin": "^8.12.1",
     "ramda": "^0.26.1",
     "typescript": "^3.8.3"
   },
