+++ src/index.ts
@@ -1,82 +1,19 @@
-import axios from 'axios'
 import { compose, isNil } from 'ramda'
 import { IncomingWebhook } from '@slack/webhook'
 
+import { getWeeklyData } from './api'
 import {
-  getFilteredPullrequest,
   generatePRInformation,
   getPersonalProjects,
   getProjectsGroupbyRepository,
   getContributionByRepository,
   calculateContributionRatio,
-  getToday,
 } from './utils'
 import { generateTotalInfoMessage, generatePersonalProjectMessage, generateContributionMessage } from './messages'
 import { weeklyMessageBlock } from './messageBlock'
+import { updateDailyData } from './firestore'
 
-const instance = axios.create({
-  baseURL: 'https://api.github.com/graphql',
-  headers: {
-    Authorization: `bearer ${process.env.GITHUB_API_KEY}`,
-  },
-})
-
-const query = (from: string, to: string) => `
-  query {
-    search(query: "author:stardustrain created:${from}..${to}", type: ISSUE, first:100) {
-      issueCount
-      nodes {
-        __typename
-        ... on PullRequest {
-          createdAt
-          closed
-          title
-          additions
-          deletions
-          state
-          author {
-            login
-          }
-          commits {
-            totalCount
-          }
-          repository {
-            name
-            owner {
-              login
-            }
-          }
-        }
-      }
-    }
-  }
-`
-
-const getWeeklyData = async () => {
-  try {
-    const today = getToday()
-    const from = today.subtract(6, 'day').format('YYYY-MM-DD')
-    const to = today.subtract(1, 'day').format('YYYY-MM-DD')
-    const {
-      data: {
-        data: {
-          search: { nodes },
-        },
-      },
-    } = await instance.post<GithubResponse>('', {
-      query: query(from, to),
-    })
-    return {
-      nodes: getFilteredPullrequest(nodes),
-      from,
-      to,
-    }
-  } catch (e) {
-    console.error(e)
-  }
-}
-
-const generateData = async () => {
+const generateMessageBlock = async () => {
   try {
     const res = await getWeeklyData()
     if (isNil(res)) {
@@ -109,7 +46,7 @@ export const sendGithubWeeklyReportWebhook = async () => {
   }
 
   const webhook = new IncomingWebhook(url)
-  const messageBlock = await generateData()
+  const messageBlock = await generateMessageBlock()
 
   if (isNil(messageBlock)) {
     throw Error()
@@ -122,3 +59,5 @@ export const sendGithubWeeklyReportWebhook = async () => {
     console.error(e)
   }
 }
+
+updateDailyData()
